script de nettoyage des structures inactives v4

+++ src/tools/scripts/archived/remove-user-structure-inactif.js
@@ -2,10 +2,14 @@
 'use strict';
 
 const { execute } = require('../../utils');
+const { program } = require('commander');
 const whiteList = ['EXAMEN_COMPLEMENTAIRE_COSELEC', 'REFUS_COSELEC', 'ABANDON', 'ANNULEE', 'DOUBLON', 'NEGATIF'];
 
 execute(__filename, async ({ db, logger, exit, Sentry }) => {
   try {
+    program.option('-f, --fix', 'fix: correction en base)');
+    program.parse(process.argv);
+    const fix = program.fix;
     // Recupération des structures dont le statut est en 'ABANDON' ou 'ANNULEE'.
     const structures = await db.collection('structures').find({
       'statut': { $in: whiteList }
@@ -35,15 +39,21 @@ execute(__filename, async ({ db, logger, exit, Sentry }) => {
       
       // Nous supprimons les mises en relation à supprimer.
       if (misesEnRelationASupprimer.length > 0) {
-        await db.collection('misesEnRelation').deleteMany({ _id: 'structure.$id' });
-        logger.info(`Nombre de mises en relation à supprimer: ${misesEnRelationASupprimer.length}`);
+        if (fix) {
+          await db.collection('misesEnRelation').deleteMany({ _id: 'structure.$id' });
+        }
+        logger.info(`Id structure ${structure.idPG} - Nombre de mises en relation à supprimer: ${misesEnRelationASupprimer.length}`);
       }
       // Nous supprimons les utilisateurs associés à la structure.
-      const deleteResult = await db.collection('users').deleteMany({ 'entity.$id': structure._id });
-
+      const nbUsers = await db.collection('users').countDocuments({ 'entity.$id': structure._id });
+      
       // Après avoir supprimé les utilisateurs, nous mettons à jour le champ 'createdUser' de la structure.
-      if (deleteResult.deletedCount > 0) {
-        await db.collection('structures').updateOne({ _id: structure._id }, { $set: { userCreated: false } });
+      if (nbUsers > 0) {
+        if (fix) {
+          await db.collection('users').deleteMany({ 'entity.$id': structure._id });
+          await db.collection('structures').updateOne({ _id: structure._id }, { $set: { userCreated: false } });
+        }
+        logger.info(`Id structure ${structure.idPG} - Nombre d'utilisateurs supprimés: ${nbUsers}`);
       }
     }
     logger.info('Fin de suppression des structures inactives');
